requirements.txt
Flask==3.0.3
Pillow==10.4.0
waitress==3.0.0
python-dotenv>=1.0.0
